--- shared/pages/NewWallet/NewWallet.scss
@@ -1,262 +1,0 @@
-.newWallet {
-    margin: 30px 0;
-    font-family: 'Manrope-Bold',sans-serif;
-}
-
-.newWalletHeading {
-    font-weight: 800;
-    font-size: 40px;
-    line-height: 120%;
-    color: #1F2D48;
-    margin-bottom: 30px;
-}
-
-.newWalletTable {
-    width: 100%;
-    td:first-child {
-        padding-left: 0;
-    }
-}
-  
-
-.newWalletNav {
-    padding: 0;
-    margin-top: 60px;
-    margin-bottom: 0px;
-}
-
-.newWalletNavItem {
-    display: inline-block;
-    margin-right: 40px;
-}
-
-.newWalletNavItem.active {
-    a {
-        border-bottom: 3px solid #6144E5;
-        color: #1F2D48!important;
-    }
-}
-
-.newWalletNavItemLink {
-    cursor: pointer;
-    font-size: 15px;
-    line-height: 140%;
-    padding-bottom: 10px;
-    color: #8E9AA3!important;
-}
-
-.newWalletBalance {
-    margin-top: 40px;
-    width: 31%;
-}
-
-.yourAssets,
-.activity {
-   display: none;
-}
-
-.yourAssets.active,
-.activity.active {
-    display: block;
-}
-
-.newWalletContent {
-    display: flex;
-    justify-content: space-between;
-}
-
-.yourAssets,
-.activity {
-    background: #FFFFFF;
-    box-shadow: 0px 4px 16px rgba(0, 0, 0, 0.08), 0px 1px 4px rgba(0, 0, 0, 0.08);
-    border-radius: 4px;
-    padding: 25px;
-    margin-top: 40px;
-    width: 66%;
-}
-
-.yourAssetsHeading,
-.activityHeading{
-    color: #1F2D48;
-    font-size: 24px;
-    line-height: 33px;
-    font-weight: bold;
-    margin-bottom: 10px;
-}
-
-.yourAssetsDescr {
-    color: #8E9AA3;
-    font-size: 13px;
-    font-family: 'Manrope-Medium',sans-serif;
-    line-height: 150%;
-}
-
-.yourAssetsTable {
-    margin-top: 27px;
-}
-
-.yourBalance {
-    background: #FFFFFF;
-    box-shadow: 0px 4px 16px rgba(0, 0, 0, 0.08), 0px 1px 4px rgba(0, 0, 0, 0.08);
-    border-radius: 4px;
-    height: fit-content;
-}
-
-.yourBalanceTop {
-    padding: 25px;
-    position: relative;
-}
-.yourBalanceDescr {
-    color: #8E9AA3;
-    font-size: 13px;
-    font-weight: bold;
-}
-.yourBalanceValue {
-    display: flex;
-    align-items: end;
-    width: max-content;
-    justify-content: space-between;   
-    img {
-        margin-top: 10px;
-        width: 10px;
-    }
-    p {
-        font-size: 32px;
-        line-height: 44px;
-        color: #1F2D48;
-        font-weight: bold;
-        margin: 0 10px;
-    }
-    span {
-        font-size: 18px;
-        line-height: 25px;
-        color: #6144E5;
-        font-weight: bold;
-        margin-top: 5px;
-    }
-}
-
-.yourBalanceBottom {
-    position: relative;
-    padding: 25px;
-    display: flex;
-    justify-content: space-between;
-    button {
-        width: 48%;
-        cursor: pointer;
-        z-index: 1;
-    }
-    &:before {
-        position: absolute;
-        width: 100%;
-        height: 100px;
-        left: 0;
-        bottom: 0;
-        opacity: 0.1;
-        content: '';
-        background: linear-gradient(180deg, #6144E5 0%, rgba(97, 68, 229, 0) 100%);
-    }
-    height: 95px;
-}
-
-.yourBalanceCurrencies {
-    color: rgba(0, 0, 0, 0.1);
-    position: absolute;
-    background-color: white;
-    right: 25px;
-    top: 25px;
-    width: 45px;
-    justify-content: space-between;
-    display: flex;
-    span {
-        width: 1px;
-        height: 16px;
-        background: rgba(0, 0, 0, 0.1);
-        display: inline-block;
-        margin-top: 6px;
-    }
-    button {
-        padding: 0;
-        cursor: pointer;
-    }
-    button.active {
-        filter: brightness(0);
-    }
-}
-
-@media all and (min-width: 993px) {
-    .newWalletTable {
-        display: inline-table;
-    }
-}
-
-@media all and (min-width: 768px) and (max-width: 992px) {
-    .yourBalance {
-        width: 45%;
-    }
-    .yourAssets {
-        width: 53%;
-    }
-    .activity {
-        width: 100%;
-    }
-    .newWalletTable {
-        display: inline-table;
-        td:first-child {
-            padding-left: 0px;
-        }
-    }
-    .yourBalance {
-        display: none;
-    }
-    .yourBalance.active {
-        display: block;
-    }
-}
-
-
-@media all and (min-width: 320px) and (max-width: 767px) {
-    .newWalletContent {
-        flex-direction: column;
-        margin: 0px -20px;
-    }
-    .yourBalance,
-    .yourAssets,
-    .activity {
-        box-shadow: none;
-        width: 100%;
-    }
-    .newWalletTable {
-        display: inline-table;
-        td:first-child {
-            padding-left: 25px;
-            left: auto;
-        }
-    }
-    .yourBalanceBottom {
-        &:before {
-            display: none;
-        }
-    }
-    .yourAssets {
-        padding: 0;
-        button {
-            width: calc(100% - 65px);
-            margin: auto;
-        }
-    }
-    .yourBalance,
-    .yourBalanceCurrencies {
-        background: rgb(253, 253, 253);
-    }
-    .yourAssetsHeading,
-    .yourAssetsDescr {
-        padding: 0 25px;
-    }
-    .yourBalance {
-        display: none;
-    }
-    .yourBalance.active {
-        display: block;
-    }
-}
