+++ shared/pages/OldWallet/Wallet.scss
@@ -1,0 +1,65 @@
+.coinsActionsBlock {
+  margin: 25px 0px;
+}
+
+.sectionWalletMobile {
+  margin-top: -88px;
+  &_widget {
+    margin-top: 0;
+  }
+}
+
+.sectionWallet {
+  margin-top: 0px;
+  &_widget {
+    margin-top: 0;
+  }
+}
+.inform {
+  color: #626262;
+  margin-top: 40px;
+}
+
+.informHeading {
+  margin-bottom: 30px;
+}
+
+.pageLine {
+  padding-bottom: 25px;
+  &_widget {
+    padding-top: 0;
+  }
+}
+.depositText {
+  margin-bottom: -15px;
+  display: block;
+}
+
+@media all and (min-width: 570px) and (max-width:700px) {
+	.sectionWalletMobile {
+		margin-top: -88px;
+	}
+}
+
+@include media-tablet {
+  .depositText {
+    margin-bottom: 10px;
+  }
+}
+
+@include media-mobile {
+	.sectionWalletMobile {
+		margin-top: 0px;
+    margin-bottom: -60px;
+	}
+  .depositText {
+    margin-bottom: 0px;
+  }
+  .pageLine {
+    height: auto;
+    padding-bottom: 0;
+  }
+  .inform {
+    margin-top: 10px;
+  }
+}
