+++ shared/pages/Wallet/components/NotityBlock/NotifyBlock.scss
