+++ shared/pages/Wallet/components/NotityBlock/NotifyBock.js
